move modifiers into spawnable module
