+++ thetawave_lib/src/spawn/resources/spawner.rs
@@ -2,9 +2,8 @@ use crate::{
     constants::{ARENA_MAX_X, ARENA_MAX_Y, ARENA_MIN_X, ARENA_SPAWN_OFFSET, SPAWNER_Y_OFFSET},
     entities::SpawnableType,
     resources::{InvasionFormationPool, InvasionRandomPool, SpriteSheetsResource},
-    spawnable::{
-        entities::spawn_spawnable,
-        resources::{ConsumablesResource, EffectsResource, ItemsResource, MobsResource},
+    spawnable::resources::{
+        spawn_spawnable, ConsumablesResource, EffectsResource, ItemsResource, MobsResource,
     },
 };
 use amethyst::{
