+++ thetawave_lib/src/spawn/systems/despawn.rs
@@ -15,6 +15,7 @@ use amethyst::{
 pub struct DespawnAtBorderSystem;
 
 impl<'s> System<'s> for DespawnAtBorderSystem {
+    /// Data used by the system
     type SystemData = (
         Entities<'s>,
         ReadStorage<'s, DespawnAtBorderComponent>,
@@ -23,6 +24,7 @@ impl<'s> System<'s> for DespawnAtBorderSystem {
         Write<'s, EventChannel<MobReachedBottomEvent>>,
     );
 
+    /// System game logic
     fn run(
         &mut self,
         (entities, despawn_borders, transforms, mobs, mut mob_reached_bottom_event_channel): Self::SystemData,
@@ -65,12 +67,14 @@ impl<'s> System<'s> for DespawnAtBorderSystem {
 pub struct DespawnTimeLimitSystem;
 
 impl<'s> System<'s> for DespawnTimeLimitSystem {
+    /// Data used by the system
     type SystemData = (
         Entities<'s>,
         WriteStorage<'s, DespawnTimeLimitComponent>,
         Read<'s, Time>,
     );
 
+    /// System game logic
     fn run(&mut self, (entities, mut timelimits, time): Self::SystemData) {
         for (timed_entity, time_component) in (&*entities, &mut timelimits).join() {
             if time_component.duration > 0.0 {
