--- thetawave_lib/src/spawnable/entities/mod.rs
@@ -1,4 +1,0 @@
-mod blast;
-mod spawn;
-
-pub use self::{blast::spawn_blasts, spawn::spawn_spawnable};
