+++ thetawave_lib/src/spawnable/systems/mob.rs
@@ -49,10 +49,12 @@ impl<'s> System<'s> for MobBehaviorSystem {
 /// Handles destruction of mob
 #[derive(Default)]
 pub struct MobDestroyedSystem {
+    /// Reads from the mob destroyed event channel
     event_reader: Option<ReaderId<MobDestroyedEvent>>,
 }
 
 impl<'s> System<'s> for MobDestroyedSystem {
+    /// Data used by the system
     type SystemData = (
         Read<'s, EventChannel<MobDestroyedEvent>>,
         Entities<'s>,
@@ -69,6 +71,7 @@ impl<'s> System<'s> for MobDestroyedSystem {
         ReadExpect<'s, Sounds>,
     );
 
+    /// Sets up event readers
     fn setup(&mut self, world: &mut World) {
         Self::SystemData::setup(world);
         self.event_reader = Some(
@@ -78,6 +81,7 @@ impl<'s> System<'s> for MobDestroyedSystem {
         );
     }
 
+    /// System game logic
     fn run(
         &mut self,
         (
