+++ thetawave_lib/src/spawnable/systems/mod.rs
@@ -1,3 +1,5 @@
+//! Systems for spawnable entities
+
 mod mob;
 mod modifiers;
