+++ thetawave_lib/src/weapons/systems/manualblaster_system.rs
@@ -8,8 +8,10 @@ use amethyst::{
 pub struct ManualBlasterSystem;
 
 impl<'s> System<'s> for ManualBlasterSystem {
+    /// Data used by the system
     type SystemData = (Read<'s, Time>, WriteStorage<'s, ManualFireComponent>);
 
+    /// System game logic
     fn run(&mut self, (time, mut manual_fires): Self::SystemData) {
         for manual_fire in (&mut manual_fires).join() {
             if manual_fire.timer > 0.0 && !manual_fire.ready {
